add code

+++ GroupKeyIndex2/types.h
@@ -1,7 +1,8 @@
 #ifndef TYPES_H_
 #define TYPES_H_
 #include <stdexcept>
+#include <sys/types.h>
 
-typedef uint32_t value_id_t;
+typedef size_t value_id_t;
 
 #endif
